[navigator plugin] Navigator Plugin Client
This is a first iteration of an enriched Navigator client library
used for creating custom metadata entities and relationships.
This includes a set of Entity and Relation definitions that make
up the default plugin-model, a plugin-client module which contains
classes that communicate with Navigator, and 2 small examples
showing how to use this library to set tags on HDFS directories and
files, creating custom entities, and linking custom entities with
existing Navigator entities using custom defined relations.