[model][client] Custom JSON serialization and change annotation targets
This is a fairly major refactor of the plugin client. The
annotations now target fields instead of methods in order to make
inheritance easier and the code more readable. Next, instead of
creating intermediate Map objects to represent the JSON data to
be written, we use a custom JSON serializer that uses the field
annotations to figure out what needs to be written.

In order to make testing a little easier, this commit also changes
the MetadataWriter to interact with an OutputStream instead of a
Writer.