[model][client] Add model name to MClass and change request format
We need the model name to establish a correspondence between the metadata
json object and the server-side models upon deserialization. This model name
is a mandatory attribute of @MClass. Relation objects no longer need to be
annotated with @MClass because the model correspondence is established via
the type attribute. Finally, instead of a comingled list of entities and
relations, the new request format is an object with 'entities' and 'relations'
fields and each field value is an array.