[examples] Small refactoring to enable examples
Renamed directories to remove the "plugin-" prefix. The prefix remains for the artifact id's
so that it is easier to distinguish in Maven imports.
Change order of c'tor arguments for HdfsEntity and modify examples to conform to entity type
now being required.
Change file format and compression type to be enums instead of strings to conform to server
side constraints