[client] Add configuration to control autocommit behavior (Closes #24)
By default the Navigator server throttles the rate at which it makes
new metadata from the SDK available. For testing and development, the
user can set a configuration variable autocommit=true to force the
server to perform a full commit immediately so there's no lag from
running the client API call to when the metadata is searchable in
Navigator. This is NOT recommended for production use.