[model][client] Custom metadata model registration
To support type custom fields and richer metadata modeling, we implement
the registerModels feature of the SDK to convert @MClass models into JSON
to be registered with the server. We're using the new custom property feature
in Navigator 2.6 to create managed custom properties that have a type and
allow validation constraints to be specified. These properties belong to namespaces
and are associated with various meta-classes (i.e., hdfs entities,
Hive tables, etc).