SDk Support for CDH 5.10 and above
From this version of the SDK, there will no longer be any ID generators
within the SDK. The user will either have to pass in the ID(if the user knows
the correct id of an entity) or the user can also pass in the required parameters
for each type of entity and the server would generate the identity based on that.
For a given Entity, the server would generate the same identity always for the
same parameters passed to it.