[client] Adding validation for entities and relations
Added validation checks for an Entity. For relations added validation
checks where either the id or the attirbutes required to generate the id
are present in the relation. Ran all the examples except the set tags for Hive.
We will be supporting Hive in the next version. Have also added unit tests for Entity